Moved OrderScratchPad to presentational
