+++ client/src/components/Presentational/OrderScratchPad/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { flexCenterMixin } from '../../global-styles/mixins';
+import { flexCenterMixin } from '../../../global-styles/mixins';
 
 export const Container = styled.div`
   display: flex;
